+++ packages/backend/api/src/router.ts
@@ -8,6 +8,7 @@ import followUpsRouter from './subrouters/follow-ups-router';
 import postsRouter from './subrouters/posts-router';
 import suggestionRouter from './subrouters/suggestions-router';
 import tagsRouter from './subrouters/tags-router';
+import usersRouter from './subrouters/users-router';
 
 const router = express.Router();
 
@@ -22,6 +23,7 @@ router.use('/follows', followUpsRouter);
 router.use('/posts', postsRouter);
 router.use('/suggestion', suggestionRouter);
 router.use('/tags', tagsRouter);
+router.use('/users', usersRouter);
 
 router.use('*', (req, res) => {
   res.status(404).json({ message: `Resource ${req.path} not found` });
